No need to have separate websocketFeedServer
